Rename: 클래스 파일들의 패키지 재구성
user라는 패키지에 모든 클래스를 넣고 관리했었는데, 각 클래스의 역할이 한 눈에 들어오지 않았습니다.
클래스들의 역할이 한 눈에 들어오게 패키지를 재구성하고 클래스를 배치했습니다.
