+++ src/main/java/kr/codesqaud/cafe/controller/UserController.java
@@ -1,5 +1,8 @@
-package kr.codesqaud.cafe.user;
+package kr.codesqaud.cafe.controller;
 
+import kr.codesqaud.cafe.domain.User;
+import kr.codesqaud.cafe.domain.UserForm;
+import kr.codesqaud.cafe.service.UserService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
